requirements.txt
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorful==0.5.5
distlib==0.3.7
exceptiongroup==1.1.3
fastapi==0.103.0
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
gpustat==1.1.1
grpcio==1.57.0
h11==0.14.0
hydra-core==1.3.2
idna==3.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
msgpack==1.0.5
multidict==6.0.4
numpy==1.25.2
nvidia-ml-py==12.535.77
omegaconf==2.3.0
opencensus==0.11.2
opencensus-context==0.1.3
packaging==23.1
pandas==2.0.3
platformdirs==3.10.0
prometheus-client==0.17.1
protobuf==4.24.2
psutil==5.9.5
py-spy==0.3.14
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
ray==2.6.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
rsa==4.9
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
starlette==0.27.0
tensorboardX==2.6.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
virtualenv==20.21.0
wcwidth==0.2.6
yarl==1.9.2
